move loop into public api
